python/pyproject.toml
[project]
version = "4.0.1"
name = "simwrapper"
description = "Run SimWrapper from Python"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.11"
dependencies = [
    "blosc>=1.11.2",
    "build>=1.2.2.post1",
    "click>=8.0",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "flask-restful>=0.3.10",
    "gunicorn>=23.0.0",
    "openmatrix>=0.3.5.0",
    "pypatch>=1.0.2",
    "pyyaml>=6.0.2",
    "tables>=3.10.2",
    "waitress>=3.0.2",
]

[project.scripts]
simwrapper = "simwrapper.cli:cli"
simwrapper-patch-windows = "patches.apply_pytables_patch:apply_patches"

[project.urls]
Homepage = "https://docs.simwrapper.app"
Repository = "https://github.com/simwrapper/simwrapper-python-tools"

[build-system]
requires = ["hatchling", "build"]
build-backend = "hatchling.build"
